Move restore_cache_state() to call_frame.rs
