+++ crates/vm/levm/src/execution_handlers.rs
@@ -30,7 +30,7 @@ impl<'a> VM<'a> {
                     return Err(error);
                 }
 
-                self.restore_state(backup, current_call_frame.call_frame_backup.clone())?;
+                self.restore_state(backup, current_call_frame)?;
 
                 Ok(ExecutionReport {
                     result: TxResult::Revert(error),
@@ -202,7 +202,7 @@ impl<'a> VM<'a> {
                 Err(error) => {
                     // Revert if error
                     current_call_frame.gas_used = current_call_frame.gas_limit;
-                    self.restore_state(backup, current_call_frame.call_frame_backup.clone())?;
+                    self.restore_state(backup, current_call_frame)?;
 
                     return Ok(ExecutionReport {
                         result: TxResult::Revert(error),
@@ -249,7 +249,7 @@ impl<'a> VM<'a> {
         let output = std::mem::take(&mut current_call_frame.output); // Bytes::new() if error is not RevertOpcode
         let gas_used = current_call_frame.gas_used;
 
-        self.restore_state(backup, current_call_frame.call_frame_backup.clone())?;
+        self.restore_state(backup, current_call_frame)?;
 
         Ok(ExecutionReport {
             result: TxResult::Revert(error),
