+++ crates/vm/levm/src/vm.rs
@@ -360,9 +360,9 @@ impl<'a> VM<'a> {
     pub fn restore_state(
         &mut self,
         backup: StateBackup,
-        call_frame_backup: CallFrameBackup,
+        current_call_frame: &mut CallFrame,
     ) -> Result<(), VMError> {
-        self.restore_cache_state(call_frame_backup)?;
+        current_call_frame.restore_cache_state(&mut self.db)?;
         self.accrued_substate = backup.substate;
         self.env.refunded_gas = backup.refunded_gas;
         self.env.transient_storage = backup.transient_storage;
@@ -386,7 +386,12 @@ impl<'a> VM<'a> {
     pub fn execute(&mut self) -> Result<ExecutionReport, VMError> {
         if let Err(e) = self.prepare_execution() {
             // We need to do a cleanup of the cache so that it doesn't interfere with next transaction's execution
-            self.restore_cache_state(self.current_call_frame()?.call_frame_backup.clone())?;
+            let mut current_call_frame = self
+                .call_frames
+                .pop()
+                .ok_or(VMError::Internal(InternalError::CouldNotPopCallframe))?;
+            current_call_frame.restore_cache_state(&mut self.db)?;
+            self.call_frames.push(current_call_frame);
             return Err(e);
         }
